+++ data-raw/STARdata.R
@@ -1,6 +1,7 @@
 library(AER)
 data(STAR)
 
-STARdata <- STAR
+STARdata <-
+    data.frame(studentid=rownames(STAR), STAR, row.names=NULL)
 
 usethis::use_data(STARdata, overwrite = TRUE)